Don't follow symlinks when deleting directories
The existing commons-io implementation follows symlinks to recursively
delete files and I can't see how that would be desired behaviour. It does
currently create issues on some of our build-agents.

Switching to using walkFileTree/delete in java.nio.Files let's you specify
how to treat symlinks and defaults to not following them. Thus we can delete
the link and not the linked directory's content.

As an aside, this is the same symlink behaviour that `git clean -dff` has.

FileUtilTest didn't contain tests for the method I changed so I rather wrote a
test in the place that was relevant to the specific issue (DirectoryCleanerTest)

(Will likely fix https://github.com/gocd/gocd/issues/714 as well)