Wrap calls to FileUtils.deleteQuietly
On closer inspection the main issue with symlinks
and commons-io seems to be that `FileUtils.deleteQuietly`
doesn't check whether the initial argument is a symlink.
Upon recursing into the intial directory, subdirectories
get deleted by `forceDelete`, which in turn calls
`deleteDirectory` which does check symlinks.

This change swaps out calls to `FileUtils.deleteQuietly`
with a local `FileUtil.deleteQuietly` method that calls
`FileUtils.forceDelete` directly and wraps the IOException
to behave the same as `deleteQuietly`

The previous change to use nio.Files was reverted in favour
of consistency and `deleteFolder` now just points to the local
`deleteQuietly`.